pyproject.toml
[tool.poetry]
name = "knowledge-plus"
version = "0.1.0"
description = ""
authors = ["Mikhail Gubin <mgubin82@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.1"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
python-dotenv = "^1.1.1"
ipython = "^9.4.0"
coverage = "^7.9.2"


[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
